webchat/requirements.txt

aiohttp==3.8.1
botbuilder-core==4.14.1
Flask==2.2.2
